Add reproduction of mutation error
Unlike `Query` errors, Mutation errors throw an actual error, which
does not show up in the `error` render prop within tests.

In practice this is rendered properly outside of test, but Jest will fail because an error is thrown.

The only way (currently) to match the real world behavior is to set an
`onError` handler on the `Mutation` component, or to set an error
policy for mutation on the top level client:

```javascript
const client = new ApolloClient({
  defaultOptions: {
    mutate: {
      errorPolicy: 'all'
    }
  },
});
```